--- src/main/java/seoham/seohamspring/domain/User.java
@@ -1,17 +1,0 @@
-package seoham.seohamspring.domain;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long userIdx;
-
-    private String email;
-
-    private String passWord;
-
-    private String nickName;
-}
