--- src/main/java/seoham/seohamspring/post/CreateUserRequest.java
@@ -1,18 +1,0 @@
-package seoham.seohamspring.domain;
-import javax.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class CreateUserRequest{
-
-    private String email;   // 유저 이메일
-
-    private String passWord;  // 유저 비밀번호
-
-    private String nickName; // 유저 닉네임
-
-}
