--- src/main/java/seoham/seohamspring/post/CreateUserResponse.java
@@ -1,12 +1,0 @@
-package seoham.seohamspring.domain;
-import javax.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-@Getter
-@Setter
-@AllArgsConstructor
-public class CreateUserResponse{
-    private long userIdx;  // 유저 Idx
-
-}
