--- src/main/java/seoham/seohamspring/service/UserService.java
@@ -1,11 +1,0 @@
-package seoham.seohamspring.service;
-
-
-import seoham.seohamspring.domain.CreateUserRequest;
-import seoham.seohamspring.domain.CreateUserResponse;
-
-public interface UserService {
-    //회원가입
-    CreateUserResponse createUser(CreateUserRequest createUserRequest);
-
-}
