+++ src/main/java/seoham/seohamspring/user/User.java
@@ -1,4 +1,17 @@
 package seoham.seohamspring.user;
 
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
 public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long userIdx;
+
+    private String email;
+
+    private String passWord;
+
+    private String nickName;
 }
