+++ src/main/java/seoham/seohamspring/user/UserController.java
@@ -1,15 +1,11 @@
-package seoham.seohamspring.controller;
+package seoham.seohamspring.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import seoham.seohamspring.domain.CreateUserRequest;
-import seoham.seohamspring.domain.CreateUserResponse;
 import seoham.seohamspring.domain.PostRequest;
 import seoham.seohamspring.repository.PostRepository;
-import seoham.seohamspring.repository.UserRepository;
 import seoham.seohamspring.service.PostService;
-import seoham.seohamspring.service.UserService;
 
 @Controller
 @RequestMapping("/user")
